+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0 (unreleased)
 
 - Updated Polars to 0.26.1
+- Added data type classes
 
 ## 0.1.5 (2022-12-22)