+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updated Polars to 0.26.1
 - Added data type classes
 - Changed `dtype` and `schema` methods to return data type class instead of symbol
+- Dropped support for Ruby < 3
 
 ## 0.1.5 (2022-12-22)